Update minimal ethereum environment

+++ minimal-ethereum-env/contracts/SimpleStorage.sol
@@ -1,20 +1,21 @@
 pragma solidity >=0.4.21 <0.7.0;
 
+
 contract SimpleStorage {
-  uint public data;
+  uint256 public data;
 
-  event Set (uint x);
+  event Set(uint256 x);
 
-  constructor(uint initVal) public {
+  constructor(uint256 initVal) public {
     data = initVal;
   }
 
-  function set(uint x) public {
+  function set(uint256 x) public {
     emit Set(x);
     data = x;
   }
 
-  function get() public view returns (uint retVal) {
+  function get() public view returns (uint256 retVal) {
     return data;
   }
 }
